denso888-backup/requirements.txt
# DENSO888 Essential Dependencies - Fixed

# Core Data Processing  
pandas>=2.0.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39

# Excel Processing
openpyxl>=3.1.0

# Configuration
python-dotenv>=1.0.0

# Progress Display
tqdm>=4.65.0

# Build Tool (Development Only)
pyinstaller>=5.10.0
